+++ CHANGELOG.md
@@ -11,6 +11,17 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ## [Released]
 
+## [1.10.4] - 2020-12-17
+
+### Added
+- database dump
+- more output for the dev mode
+
+### Fixed
+- database validation for irrelevant groups
+- db remove command options not collapsing properly
+
+
 ## [1.10.3] - 2020-11-11
 
 ### Fixed
@@ -189,6 +200,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 <!-- Versions -->
 [Unreleased]: https://github.com/RLNT/sinus-staff-list/compare/v1.0.0...HEAD
 [Released]: https://github.com/RLNT/sinus-staff-list/releases
+[1.10.4]: https://github.com/RLNT/sinus-staff-list/compare/v1.10.3..v1.10.4
 [1.10.3]: https://github.com/RLNT/sinus-staff-list/compare/v1.10.2..v1.10.3
 [1.10.2]: https://github.com/RLNT/sinus-staff-list/compare/v1.10.1..v1.10.2
 [1.10.1]: https://github.com/RLNT/sinus-staff-list/compare/v1.10.0..v1.10.1